v0.0.11 - refactor inputs
