+++ src/graphql/mod.rs
@@ -1,2 +1,3 @@
+pub mod input;
 pub mod resolver;
 pub mod schema;
