+++ src/graphql/schema/create_entities/create_resolver/create_resolver_input_value/register_inputs/mod.rs
@@ -1,0 +1,17 @@
+use async_graphql::dynamic::InputObject;
+use log::debug;
+
+use crate::graphql::schema::ServiceSchemaBuilder;
+
+impl ServiceSchemaBuilder {
+    pub fn register_inputs(mut self, inputs: Vec<InputObject>) -> Self {
+        debug!("Registering Inputs");
+
+        for input in inputs {
+            debug!("Registering Input: {}", input.type_name());
+            self.schema_builder = self.schema_builder.register(input);
+        }
+
+        self
+    }
+}
